# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/opportunities": {
      "get": {
        "operationId": "OpportunitiesController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "description": "User role",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of opportunities"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all opportunities (published for candidates, all for others)",
        "tags": [
          "Opportunities"
        ]
      },
      "post": {
        "operationId": "OpportunitiesController_create",
        "parameters": [
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpportunityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Opportunity created with status pending"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Post a new opportunity (recruiter / placement_officer)",
        "tags": [
          "Opportunities"
        ]
      }
    },
    "/opportunities/{id}": {
      "get": {
        "operationId": "OpportunitiesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunity details"
          },
          "404": {
            "description": "Opportunity not found"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get a single opportunity by ID",
        "description": "Returns full details of one opportunity including status, eligibility, and tags.",
        "tags": [
          "Opportunities"
        ]
      },
      "put": {
        "operationId": "OpportunitiesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOpportunityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update an opportunity (recruiter / placement_officer)",
        "tags": [
          "Opportunities"
        ]
      },
      "delete": {
        "operationId": "OpportunitiesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete an opportunity",
        "tags": [
          "Opportunities"
        ]
      }
    },
    "/opportunities/{id}/approve": {
      "patch": {
        "description": "Marks a pending opportunity submission as **approved**. Workflow: `pending → approved → Published`. Set `x-role: placement_officer`.",
        "operationId": "OpportunitiesController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Opportunity ID to approve",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunity status set to approved"
          },
          "404": {
            "description": "Opportunity not found"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Approve a pending opportunity (placement_officer only)",
        "tags": [
          "Opportunities"
        ]
      }
    },
    "/opportunities/{id}/publish": {
      "patch": {
        "description": "Makes an approved opportunity **live** so that candidates can view and apply. This is the final step in the review workflow: `pending → approved → Published`. Only opportunities with status `approved` should be published, but the endpoint accepts any status for flexibility.",
        "operationId": "OpportunitiesController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Opportunity ID to publish",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunity status set to Published and is now live for candidates"
          },
          "404": {
            "description": "Opportunity not found"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Publish an approved opportunity to candidates (placement_officer only)",
        "tags": [
          "Opportunities"
        ]
      }
    },
    "/opportunities/{id}/reject": {
      "patch": {
        "operationId": "OpportunitiesController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Reject an opportunity (placement_officer only)",
        "tags": [
          "Opportunities"
        ]
      }
    },
...
  },
  "info": {
    "title": "CareerNest API",
    "description": "## CareerNest Placement Portal — REST API\n\nFull-stack placement management system built with **NestJS + TypeScript**.\n\n### Authentication\nAll protected endpoints require an `x-role` header. Set it in **Authorize** (top-right) or per-request:\n\n| Role | Who uses it | Example header |\n|---|---|---|\n| `candidate` | Students applying for jobs | `x-role: candidate` |\n| `recruiter` | Company HR posting opportunities | `x-role: recruiter` |\n| `alumni` | Alumni providing referrals | `x-role: alumni` |\n| `placement_officer` | Admin managing the platform | `x-role: placement_officer` |\n\n### Demo Flow\n1. `GET /opportunities` — view live jobs\n2. `POST /applications` — candidate applies\n3. `PATCH /applications/{id}/status` — recruiter shortlists\n4. `POST /assessments` — recruiter schedules test (auto-notifies applicants)\n5. `GET /notifications?candidateId=1` — candidate sees announcement\n\n### Base URL\n`http://localhost:3000`",
    "version": "1.0",
    "contact": {
      "name": "Sai Ganesh & Shammer Basha Shaik",
      "email": "careernest@college.in"
    }
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Local Development Server" }
  ],
  "components": {
    "schemas": {
...
      "CreateOpportunityDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Software Engineer"
          },
          "company": {
            "type": "string",
            "example": "Google"
          },
          "location": {
            "type": "string",
            "example": "Bangalore, Karnataka"
          },
          "deadline": {
            "type": "string",
            "example": "2026-05-25"
          },
          "salary": {
            "type": "string",
            "example": "12 LPA"
          },
          "openings": {
            "type": "number",
            "example": 15
          },
          "cgpa": {
            "type": "number",
            "example": 7.5
          },
          "branches": {
            "example": [
              "CSE",
              "ECE",
              "IT"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "type": "string",
            "example": "Full-Time",
            "enum": [
              "Full-Time",
              "Internship"
            ]
          },
          "tags": {
            "example": [
              "JavaScript",
              "Node.js",
              "React"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "example": "Join our engineering team to build scalable applications."
          },
          "rounds": {
            "type": "string",
            "example": "Resume → Assessment → Technical → HR"
          },
          "backlogs": {
            "type": "string",
            "example": "Not allowed"
          },
          "submittedBy": {
            "type": "string",
            "example": "priya.kapoor@google.com",
            "description": "Email of the recruiter who posted — used for per-recruiter filtering in My Opportunities"
          }
        },
        "required": [
          "title",
          "company",
          "location",
          "deadline",
          "salary",
          "openings",
          "cgpa",
          "branches",
          "type"
        ]
      },
      "UpdateOpportunityDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Software Engineer"
          },
          "company": {
            "type": "string",
            "example": "TechCorp Solutions"
          },
          "location": {
            "type": "string",
            "example": "Bangalore, Karnataka"
          },
          "deadline": {
            "type": "string",
            "example": "2026-05-25"
          },
          "salary": {
            "type": "string",
            "example": "12 LPA"
          },
          "openings": {
            "type": "number",
            "example": 15
          },
          "cgpa": {
            "type": "number",
            "example": 7.5
          },
          "branches": {
            "example": [
              "CSE",
              "ECE",
              "IT"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "type": "string",
            "example": "Full-Time",
            "enum": [
              "Full-Time",
              "Internship"
            ]
          },
          "tags": {
            "example": [
              "JavaScript",
              "Node.js",
              "React"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string",
            "example": "Join our engineering team to build scalable applications."
          },
          "rounds": {
            "type": "string",
            "example": "Resume → Assessment → Technical → HR"
          },
          "backlogs": {
            "type": "string",
            "example": "Not allowed"
          },
          "status": {
            "type": "string",
            "example": "Published",
            "enum": [
              "Published",
              "Draft",
              "Closed",
              "pending",
              "approved",
              "rejected"
            ]
          },
          "rejectionRemark": {
            "type": "string",
            "example": "Eligibility criteria not met"
          },
          "applications": {
            "type": "number",
            "example": 78
          }
        }
      },
      "RejectDto": {
        "type": "object",
        "properties": {
          "remark": {
            "type": "string",
            "example": "Does not meet eligibility criteria"
          }
        },
        "required": [
          "remark"
        ]
      },
...
    },
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role of the caller. Valid values: `candidate`, `recruiter`, `alumni`, `placement_officer`"
      }
    }
  }
}