# backend/src/opportunities/opportunities.controller.ts
import {
  Controller, Get, Post, Put, Patch, Delete, Param,
  Body, ParseIntPipe, UseGuards, Headers, HttpCode, HttpStatus,
} from '@nestjs/common';
import { ApiTags, ApiOperation, ApiParam, ApiBody, ApiSecurity, ApiResponse, ApiHeader } from '@nestjs/swagger';
import { OpportunitiesService } from './opportunities.service';
import { CreateOpportunityDto } from './dto/create-opportunity.dto';
import { UpdateOpportunityDto } from './dto/update-opportunity.dto';
import { RolesGuard } from '../common/guards/roles.guard';
import { Roles } from '../common/decorators/roles.decorator';
import { IsOptional, IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

class RejectDto {
  @ApiProperty({ example: 'Does not meet eligibility criteria' })
  @IsString()
  remark!: string;
}

@ApiTags('Opportunities')
@ApiSecurity('x-role')
@UseGuards(RolesGuard)
@Controller('opportunities')
export class OpportunitiesController {
  constructor(private readonly service: OpportunitiesService) {}

  @Get()
  @ApiOperation({ summary: 'Get all opportunities (published for candidates, all for others)' })
  @ApiHeader({ name: 'x-role', description: 'User role', required: true })
  @ApiResponse({ status: 200, description: 'List of opportunities' })
  findAll(@Headers('x-role') role: string) {
    if (role === 'candidate') {
      return this.service.findPublished();
    }
    return this.service.findAll();
  }

  @Get(':id')
  @ApiOperation({ summary: 'Get a single opportunity by ID' })
  @ApiParam({ name: 'id', type: Number })
  findOne(@Param('id', ParseIntPipe) id: number) {
    return this.service.findOne(id);
  }

  @Post()
  @Roles('recruiter', 'placement_officer')
  @ApiOperation({ summary: 'Post a new opportunity (recruiter / placement_officer)' })
  @ApiBody({ type: CreateOpportunityDto })
  @ApiResponse({ status: 201, description: 'Opportunity created with status pending' })
  create(@Body() dto: CreateOpportunityDto, @Headers('x-role') role: string) {
    return this.service.create(dto, role);
  }

  @Put(':id')
  @Roles('recruiter', 'placement_officer')
  @ApiOperation({ summary: 'Update an opportunity (recruiter / placement_officer)' })
  @ApiParam({ name: 'id', type: Number })
  @ApiBody({ type: UpdateOpportunityDto })
  update(@Param('id', ParseIntPipe) id: number, @Body() dto: UpdateOpportunityDto) {
    return this.service.update(id, dto);
  }

  @Patch(':id/approve')
  @Roles('placement_officer')
  @ApiOperation({
    summary: 'Approve a pending opportunity (placement_officer only)',
    description: 'Marks a pending opportunity submission as **approved** (reviewed and accepted). ' +
      'After approval, the placement officer can publish it to make it live for candidates via `PATCH /opportunities/:id/publish`.',
  })
  @ApiParam({ name: 'id', type: Number, description: 'Opportunity ID to approve' })
  @ApiResponse({ status: 200, description: 'Opportunity status set to approved' })
  @ApiResponse({ status: 404, description: 'Opportunity not found' })
  approve(@Param('id', ParseIntPipe) id: number) {
    return this.service.approve(id);
  }

  @Patch(':id/publish')
  @Roles('placement_officer')
  @ApiOperation({
    summary: 'Publish an approved opportunity to candidates (placement_officer only)',
    description:
      'Makes an approved opportunity **live** so that candidates can view and apply. ' +
      'This is the final step in the review workflow: `pending → approved → Published`. ' +
      'Only opportunities with status `approved` should be published, but the endpoint ' +
      'accepts any status for flexibility.',
  })
  @ApiParam({ name: 'id', type: Number, description: 'Opportunity ID to publish' })
  @ApiResponse({ status: 200, description: 'Opportunity status set to Published and is now live for candidates' })
  @ApiResponse({ status: 404, description: 'Opportunity not found' })
  publish(@Param('id', ParseIntPipe) id: number) {
    return this.service.publish(id);
  }

  @Patch(':id/reject')
  @Roles('placement_officer')
  @ApiOperation({ summary: 'Reject an opportunity (placement_officer only)' })
  @ApiParam({ name: 'id', type: Number })
  @ApiBody({ type: RejectDto })
  reject(@Param('id', ParseIntPipe) id: number, @Body() body: RejectDto) {
    return this.service.reject(id, body.remark);
  }

  @Delete(':id')
  @Roles('recruiter', 'placement_officer')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Delete an opportunity' })
  @ApiParam({ name: 'id', type: Number })
  remove(@Param('id', ParseIntPipe) id: number) {
    return this.service.remove(id);
  }
}